repo cleanup: [removed GraphQL implentation, set indexed imports/exports for models and abi, moved proposals models to src/models]
