--- src/data/network/graph/graphQLAPIMapper.js
@@ -1,12 +1,0 @@
-/* eslint-disable no-unused-vars */
-/* eslint-disable class-methods-use-this */
-class GraphQLAPIMapper {
-  mapAssets(rawAssets) {}
-  mapProposals(rawProposals) {}
-  mapVotes(rawVotes) {}
-  mapVouchers(rawVouchers) {}
-  mapWeavrWhitelist(rawWeavrWhitelistRecords) {}
-  // mapTokenInfo(rawWeavrTokenInfo) {}
-}
-
-export default GraphQLAPIMapper
