--- src/data/network/graph/implementation/participants/index.js
@@ -1,13 +1,0 @@
-class Partcipant {
-  constructor(address, type) {
-    this.address = address,
-    this.type = type
-  }
-}
-
-export function mapParticipants(rawParticipants) {
-  return rawParticipants.map(({
-    address,
-    type
-  }) => new Partcipant(address, type))
-}
