--- src/models/marketplace/thread.js
@@ -1,39 +1,0 @@
-/**
- * Asset model.
- * @property {string} id ID of the asset
- * @property {string} dataURI Location of the off-chain data of the asset
- * @property {string} contract Address of the asset contract
- * @property {string} symbol ERC-20 Symbol of the asset
- * @property {number} numOfShares Total supply of the asset shares
- * @property {Map<string, number>} owners The amount of shares held by all the owners
- * @property {MarketOrder[]} marketOrders The most recent state of the asset's order book
- * @property {Proposal[]} proposals Proposals made in the asset's DAO
- * @property {string} address Address of the property
- * @property {number} area Area of the property
- * @property {string} coverPictureURI Location of the cover picture data
- * @property {number} currentRent Current rent the property is making
- * @property {string} description Free-format desctiption attached to the property
- * @property {number} grossYieldPct Gross yield % of the property
- * @property {number} marketValue Market value of the property
- * @property {number} bedroomCount Number of bedrooms in the property
- * @property {number} bathroomCount Number of bathrooms in the property  
- * @property {number} yearBuilt The year property was built
- */
-class Thread {
-  constructor(
-    id,
-    variant,
-    governor,
-    erc20,
-    descriptor,
-  ) {
-    this.id = id
-    this.variant = variant
-    this.governor = governor
-    this.erc20 = erc20
-    this.descriptor = descriptor
-    
-  }
-}
-  
-export default Thread
