--- src/models/proposals/descriptorChangeProposal.js
@@ -1,48 +1,0 @@
-import BaseProposal from "./baseProposal";
-
-/**
- * Proposal model.  
-*/
-
-export default class DescriptorChangeProposal extends BaseProposal{
-
-  // Default attributes that define the "empty" state.
-  // eslint-disable-next-line class-methods-use-this
-  defaults() {
-    return {
-      id: null,
-      thread: [],
-      descriptor: ""
-    }
-  }
-
-  // constructor(
-  //   id,
-  //   thread,
-  //   frabric,
-  //   creator,
-  //   type,
-  //   state,
-  //   votes,
-  //   supermajority,
-  //   startTimestamp,
-  //   endTimestamp,
-  //   description,
-  //   descriptor
-  // ) {
-  //   super(
-  //     id,
-  //     thread,
-  //     frabric,
-  //     creator,
-  //     type,
-  //     state,
-  //     votes,
-  //     supermajority,
-  //     startTimestamp,
-  //     endTimestamp
-  //   );
-  //   this.description = description;
-  //   this.descriptor = descriptor;
-  // }
-}
