--- src/models/proposals/frabricChangeProposal.js
@@ -1,48 +1,0 @@
-import BaseProposal from "./baseProposal";
-
-/**
- * Frabric Change Proposal model.
-*/
-
-export default class FrabricChangeProposal extends BaseProposal{
-
-  // Default attributes that define the "empty" state.
-  // eslint-disable-next-line class-methods-use-this
-  defaults () {
-    return {
-      id: null,
-      thread: [],
-      frabric: "",
-      governor: "",
-    }
-  }
-  // constructor(
-  //   id,
-  //   thread,
-  //   frabric,
-  //   creator,
-  //   type,
-  //   state,
-  //   votes,
-  //   supermajority,
-  //   startTimestamp,
-  //   endTimestamp,
-  //   newFrabric,
-  //   governor
-  // ) {
-  //   super(
-  //     id,
-  //     thread,
-  //     frabric,
-  //     creator,
-  //     type,
-  //     state,
-  //     votes,
-  //     supermajority,
-  //     startTimestamp,
-  //     endTimestamp
-  //   );
-  //   this.newFrabric = newFrabric;
-  //   this.governor = governor;
-  // }
-}
